# src/main/java/com/dliriotech/tms/fleetservice/dto/TipoObservacionResponse.java
package com.dliriotech.tms.fleetservice.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class TipoObservacionResponse {
    private Integer id;
    private String nombre;
    private String ambito;
    private String descripcion;
    private Boolean activo;
}